Reworked use of vol() where appropriate changed vol()-> Double(vol()) to prevent overflows. Integers dealing with vol() changed to size_t
Overhauled sanitizer options: undefined only, made sanitizer compile time option PRIVATE

+++ lib/actions/ferm/invert/qphix/syssolver_linop_clover_qphix_iter_refine_w.h
@@ -435,9 +435,10 @@ namespace Chroma
       }
 #endif
 
-      int num_cb_sites = Layout::vol()/2;
-      unsigned long total_flops = (site_flops + (1320+504+1320+504+48)*mv_apps)*num_cb_sites;
-      double gflops = (double)(total_flops)/(1.0e9);
+      size_t num_cb_sites = Layout::vol()/2;
+      double total_flops = (static_cast<double>(site_flops) + static_cast<double>(1320+504+1320+504+48)
+															*static_cast<double>(mv_apps))*static_cast<double>(num_cb_sites);
+      double gflops = total_flops/1.0e9;
 
       double total_time = end - start;
       QDPIO::cout << "QPHIX_CLOVER_BICGSTAB_ITER_REFINE_SOLVER: Solver Time="<< total_time <<" (sec)  Performance=" << gflops / total_time << " GFLOPS" << std::endl;
